Crate polywrap

Re-exports

pub use polywrap_client as client;

pub use polywrap_client_builder as builder;

pub use polywrap_client_default_config as default_config;

pub use polywrap_core as core;

pub use polywrap_msgpack_serde as msgpack;

pub use polywrap_plugin as plugin;

pub use polywrap_resolvers as resolvers;

pub use polywrap_wasm as wasm;

pub use wrap_manifest_schemas as wrap_manifest;

pub use polywrap_resolver_extensions as resolver_extensions;

Modules

base_resolver

builder

cache

core

deserialize

embeds

extendable_uri_resolver

get_schemas

implementor

map

A map of String to serde_json::Value.

method

module

msgpack

package

package_resolver

package_to_wrapper_resolver

plugin

polywrap_base_resolver

polywrap_client_config

polywrap_client_config_builder

polywrap_core

polywrap_msgpack_serde

recursive_resolver

redirect_resolver

resolution_result_cache_resolver

resolver_vec

resolver_with_loop_guard

resolvers

runtime

simple_file_resolver

static_resolver

subinvoker

system_config

uri_resolver_aggregator

uri_resolver_aggregator_base

uri_resolver_wrapper

utils

validate

value

The Value enum, a loosely typed way of representing any valid JSON value.

versions

wasm

wasm_module

wasm_package

wasm_wrapper

web3_config

with_methods

wrap_manifest

wrapper

wrapper_cache_resolver

wrapper_resolver

Macros

forward_to_deserialize_any

Helper macro when implementing the Deserializer part of a new data format for Serde.

json

Construct a serde_json::Value from a JSON literal.

resolver_vec

uri

Structs

AnyDefinition

ArrayDefinition

BigInt

A big signed integer type.

BigIntWrapper

BigNumber

A big decimal type.

ByteBuf

Wrapper around Vec<u8> to serialize and deserialize efficiently.

CapabilityDefinition

CapabilityDefinitionGetImplementations

Client

PolywrapClient is a client used for interacting, loading, resolving and invoking wraps.

ClientConfig

CoreClientConfig

Configuration struct for implementors of Client. Can be built manually or through the ClientConfigBuilder

EnumDefinition

EnumRef

EnvDefinition

GenericDefinition

ImportedDefinition

ImportedEnumDefinition

ImportedEnvDefinition

ImportedModuleDefinition

ImportedModuleRef

ImportedObjectDefinition

InterfaceDefinition

InterfaceImplementedDefinition

JSONString

MapDefinition

MapKeyDefinition

MethodDefinition

MethodDefinitionEnv

ModuleDefinition

Number

Represents a JSON number, whether integer or floating point.

ObjectDefinition

ObjectRef

ParseBigIntError

ParseError

PluginModuleWithMethods

PluginPackage

PluginWrapper

PolywrapBaseResolver

PolywrapBaseResolverOptions

PropertyDefinition

ScalarDefinition

StreamDeserializer

Iterator that deserializes a stream into multiple JSON values.

SystemClientConfig

UnresolvedObjectOrEnumRef

Uri

Represents a Wrap URI.

UriRedirect

A utility struct to store a URI redirect.

Web3ClientConfig

WithComment

WithKind

WrapManifest01

WrapManifest01Abi

Enums

AnyManifest

Error

An enum that represents all possible errors the Polywrap Client can throw.

MapKeyType

PluginError

ScalarType

Value

Represents any valid JSON value.

Traits

ClientConfigBuilder

Defines a type that uses the builder pattern to build a ClientConfig.

CoreClient

CoreClientConfigBuilder

Defines a type that can build a CoreClientConfig.

Deserialize

A data structure that can be deserialized from any data format supported by Serde.

Invoker

Defines an object capable of invoking on URIs

PluginModule

Serialize

A data structure that can be serialized into any data format supported by Serde.

Functions

from_reader

Deserialize an instance of type T from an I/O stream of JSON.

from_slice

from_str

Deserialize an instance of type T from a string of JSON text.

from_value

Interpret a serde_json::Value as an instance of type T.

get_error_message

to_string

Serialize the given data structure as a String of JSON.

to_string_pretty

Serialize the given data structure as a pretty-printed String of JSON.

to_value

Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

to_vec

to_vec_pretty

Serialize the given data structure as a pretty-printed JSON byte vector.

to_writer

Serialize the given data structure as JSON into the I/O stream.

to_writer_pretty

Serialize the given data structure as pretty-printed JSON into the I/O stream.

Type Definitions

PluginMethod

Result

Alias for a Result with the error type serde_json::Error.

WrapManifest

WrapManifestAbi

Attribute Macros

plugin_impl

Derive Macros

Deserialize

Serialize