Crate polywrap::plugin

Modules

deserialize

error

get_schemas

implementor

method

module

package

utils

validate

versions

with_methods

wrapper

Macros

json

Construct a serde_json::Value from a JSON literal.

uri

Structs

AnyDefinition

ArrayDefinition

BigInt

A big signed integer type.

BigIntWrapper

BigNumber

A big decimal type.

ByteBuf

Wrapper around Vec<u8> to serialize and deserialize efficiently.

CapabilityDefinition

CapabilityDefinitionGetImplementations

CoreClientConfig

Configuration struct for implementors of Client. Can be built manually or through the ClientConfigBuilder

EnumDefinition

EnumRef

EnvDefinition

GenericDefinition

ImportedDefinition

ImportedEnumDefinition

ImportedEnvDefinition

ImportedModuleDefinition

ImportedModuleRef

ImportedObjectDefinition

InterfaceDefinition

InterfaceImplementedDefinition

JSONString

Map

An ordered map based on a B-Tree.

MapDefinition

MapKeyDefinition

MethodDefinition

MethodDefinitionEnv

ModuleDefinition

Number

Represents a JSON number, whether integer or floating point.

ObjectDefinition

ObjectRef

ParseBigIntError

ParseError

PluginModuleWithMethods

PluginPackage

PluginWrapper

PropertyDefinition

ScalarDefinition

StreamDeserializer

Iterator that deserializes a stream into multiple JSON values.

UnresolvedObjectOrEnumRef

Uri

Represents a Wrap URI.

UriRedirect

A utility struct to store a URI redirect.

WithComment

WithKind

WrapManifest01

WrapManifest01Abi

Enums

AnyManifest

MapKeyType

PluginError

ScalarType

Value

Represents any valid JSON value.

Traits

CoreClient

CoreClientConfigBuilder

Defines a type that can build a CoreClientConfig.

Invoker

Defines an object capable of invoking on URIs

PluginModule

Functions

from_reader

Deserialize an instance of type T from an I/O stream of JSON.

from_str

Deserialize an instance of type T from a string of JSON text.

from_value

Interpret a serde_json::Value as an instance of type T.

get_error_message

to_string

Serialize the given data structure as a String of JSON.

to_string_pretty

Serialize the given data structure as a pretty-printed String of JSON.

to_value

Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

to_vec

to_vec_pretty

Serialize the given data structure as a pretty-printed JSON byte vector.

to_writer

Serialize the given data structure as JSON into the I/O stream.

to_writer_pretty

Serialize the given data structure as pretty-printed JSON into the I/O stream.

Type Definitions

PluginMethod

WrapManifest

WrapManifestAbi

Attribute Macros

plugin_impl