Trait polywrap::Invoker

pub trait Invoker: Send + Sync {
    // Required methods
    fn invoke_raw(
        &self,
        uri: &Uri,
        method: &str,
        args: Option<&[u8]>,
        env: Option<&[u8]>,
        resolution_context: Option<Arc<Mutex<UriResolutionContext>>>
    ) -> Result<Vec<u8, Global>, Error>;
    fn get_implementations(&self, uri: &Uri) -> Result<Vec<Uri, Global>, Error>;
    fn get_interfaces(
        &self
    ) -> Option<HashMap<Uri, Vec<Uri, Global>, RandomState>>;
    fn get_env_by_uri(&self, uri: &Uri) -> Option<Vec<u8, Global>>;
}

Defines an object capable of invoking on URIs

Required Methods

fn invoke_raw( &self, uri: &Uri, method: &str, args: Option<&[u8]>, env: Option<&[u8]>, resolution_context: Option<Arc<Mutex<UriResolutionContext>>> ) -> Result<Vec<u8, Global>, Error>

Invokes a method on a given URI with optional arguments and environment. The method returns a Result containing either the msgpcack buffer of the response or an Error.

Arguments

uri - A reference to the Uri to invoke the method on.

method - The name of the method to invoke.

args - Optional msgpack buffer representing the arguments to the method.

env - Optional msgpack buffer representing the environment for the method.

resolution_context - Optional resolution context of invocation.

fn get_implementations(&self, uri: &Uri) -> Result<Vec<Uri, Global>, Error>

Returns a list of Uris that this invoker can handle.

Arguments

uri - A reference to the Uri to get the implementations for.

fn get_interfaces(&self) -> Option<HashMap<Uri, Vec<Uri, Global>, RandomState>>

Returns a list of interfaces that this invoker can handle.

fn get_env_by_uri(&self, uri: &Uri) -> Option<Vec<u8, Global>>

Returns the environment for a given Uri.

Arguments

uri - A reference to the Uri to get the environment for.